Refactor LayoutWithSidebars component and update imports
Renamed LayoutWithSidebars component files and removed unused props. Integrated Redux state to manage top indent dynamically and updated related SCSS styles. Adjusted import paths to reflect the new directory structure.
