+++ frontend-next-migration/src/preparedPages/Layouts/index.ts
